Add popup and tip custom duration and improve message queue
* Tracking custom popup and tip duration times server side stops weird
glitches happening when the server tries to send a popup or tip to the
client when they already have one being displayed.

* Queuing messages and sending only one raw text packet a tick with all
the messages combined stops a client glitch happening where if more
than one message is sent in a tick then the first message sent in that
tick is displayed for every other message.